# hek_analysis/docs/architecture/Unified-Healthcare-API_openapi.yaml
openapi: 3.0.3
info:
  title: Unified Healthcare API
  description: >
    Canonical contract consolidating HISO, KARO, and ERMS. See
    Unified-Healthcare-API_API-Contract-Design.md for full rationale,
    traceability, and reconciliation decisions. No version identifier
    appears in the URL/header/media-type per stakeholder decision (Section 9
    of the contract document) — version is tracked here via info.version and
    an append-only changelog only; no request/response shape changes without
    a coordinated migration.
  version: "1.1.1"
  x-changelog:
    - version: "1.0.0"
      date: "2026-07-19"
      description: Initial canonical contract, Phase 7 (API Contract Design).
    - version: "1.1.0"
      date: "2026-07-19"
      description: >
        Demographics split into 4 legacy-shaped endpoints instead of one merged
        endpoint; HISO's addInvoice/launchForm actions and static-mode form
        retrieval implemented (previously excluded as dead stubs); KARO's
        GetEncounterSummary and SaveScreeningCode implemented for real
        (previously excluded as confirmed-broken); rate limiting confirmed
        as a built, config-toggled capability.
    - version: "1.1.1"
      date: "2026-07-20"
      description: >
        Block 3 contract re-verification against the actual Day-1 build (all 42
        operations implemented, route-for-route and verb-for-verb match).
        One drift corrected here rather than left silently mismatched:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /patients/{patientId}/encounters/{encounterId}/lab-results:
    get:
      tags: [Lab and Radiology]
      summary: List lab results
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/sinceDate'
        - $ref: '#/components/parameters/untilDate'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of lab results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReportSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/{patientId}/encounters/{encounterId}/lab-results/{reportId}:
    get:
      tags: [Lab and Radiology]
      summary: Lab report detail/content
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/reportId'
      responses:
        '200':
          description: Lab report content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /patients/{patientId}/encounters/{encounterId}/radiology-results:
    get:
      tags: [Lab and Radiology]
      summary: List radiology results
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/sinceDate'
        - $ref: '#/components/parameters/untilDate'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        '200':
          description: List of radiology results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ReportSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/{patientId}/encounters/{encounterId}/radiology-results/{reportId}:
    get:
      tags: [Lab and Radiology]
      summary: Radiology report detail/content
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/reportId'
      responses:
        '200':
          description: Radiology report content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /patients/{patientId}/encounters/{encounterId}/screening-codes:
    post:
      tags: [Screening]
      summary: >
        Save a screening code result, with real authentication and persistence.
        Legacy SaveScreeningCode performed no token validation and always
        reported fake success without persisting (KARO-BR-06); implemented for
        real per stakeholder decision 2026-07-19 (FR-AUTH-04 bars reproducing
        the legacy no-auth behavior).
      parameters:
        - $ref: '#/components/parameters/patientId'
        - $ref: '#/components/parameters/encounterId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                value: { type: string }
      responses:
        '201':
          description: Screening code result saved
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Resource-scoped token per ADR-003 (patientId + encounterId + practiceId +
        structurally-determined originScope). Issued via POST /auth/token, or
        internally minted by the HISO SOAP edge adapter from an existing
        SessionGUID (ADR-004/ADR-007).
  parameters:
    patientId:
      name: patientId
      in: path
      required: true
      schema: { type: integer }
    encounterId:
      name: encounterId
      in: path
      required: true
      schema: { type: integer }
    reportId:
      name: reportId
      in: path
      required: true
      schema: { type: string }
...
    sinceDate:
      name: sinceDate
      in: query
      schema: { type: string, format: date }
    untilDate:
      name: untilDate
      in: query
      schema: { type: string, format: date }
    sortOrder:
      name: sortOrder
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: See Section 12 of the contract document. Optional; falls back to natural-key duplicate detection if omitted.
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      description: Data-minimized error shape — never contains PHI, stack traces, or internal exception detail (SRS §12.14).
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: Generic, static, non-PHI message. }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
      required: [title, status]
...
    ReportSummary:
      type: object
      properties:
        reportId: { type: string }
        type: { type: string }
        date: { type: string, format: date-time }
    ReportContent:
      type: object
      properties:
        reportId: { type: string }
        content: { type: string, description: Report text/content, RTF transcoding replaced with a documented modern format per FR-LAB-01. }
        contentEncoding: { type: string, enum: [plain, base64] }
...